[TASK] Add AccessService::hasGroupAccess()
