+++ Classes/Service/AccessService.php
@@ -1,0 +1,41 @@
+<?php
+
+declare(strict_types=1);
+
+namespace AOE\Crawler\Service;
+
+/*
+ * (c) 2020 AOE GmbH <dev@aoe.com>
+ *
+ * This file is part of the TYPO3 Crawler Extension.
+ *
+ * It is free software; you can redistribute it and/or modify it under
+ * the terms of the GNU General Public License, either version 2
+ * of the License, or any later version.
+ *
+ * For the full copyright and license information, please read the
+ * LICENSE.txt file that was distributed with this source code.
+ *
+ * The TYPO3 project - inspiring people to share!
+ */
+
+use TYPO3\CMS\Core\Utility\GeneralUtility;
+
+class AccessService
+{
+    /**
+     * @see \TYPO3\CMS\Frontend\Page\PageRepository::getMultipleGroupsWhereClause()
+     */
+    public static function hasGroupAccess(string $groupList, string $accessList): bool
+    {
+        if (empty($accessList)) {
+            return true;
+        }
+        foreach (explode(',', $groupList) as $groupUid) {
+            if (GeneralUtility::inList($accessList, $groupUid)) {
+                return true;
+            }
+        }
+        return false;
+    }
+}
